server/requirements.txt
numpy==1.15.2
flask==1.0.2
python-dotenv==0.9.1
keras==2.2.4
tensorflow